% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxPot.R
\name{approxPot1D}
\alias{approxPot1D}
\title{Approximate potential in 1D}
\usage{
approxPot1D(f, xs, V0 = "auto")
}
\arguments{
\item{f}{Dynamics}

\item{xs}{Vector of positions to evaluate}

\item{V0}{(Optional) Value of V at first element of xs. When default, the global minimum is assigned 0}
}
\value{
The potential estimated at each point in xs
}
\description{
Approximate potential in 1D
}
\examples{
# Flow
f = function(x) { sin(x) }

# Sampling points
xs <- seq(0, 2*pi, length.out = 1e3)

# Approximated potential
Vs <- approxPot1D(f, xs)
}
