% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxPot.R
\name{approxPot2D}
\alias{approxPot2D}
\title{Approximate potential in 1D}
\usage{
approxPot2D(f, xs, ys, V0 = 0, mode = "horizontal")
}
\arguments{
\item{f}{Dynamics}

\item{xs}{Vector xs positions to evaluate}

\item{ys}{Vector of ys positions to evaluate}

\item{V0}{(Optional) Value of V at first element of (xs,ys) (0 by default)}

\item{mode}{(Optional) Integration mode. Options are horizontal (default), vertical and mixed}
}
\value{
The potential estimated at each point (xs, ys)
}
\description{
Approximate potential in 1D
}
\examples{
# Flow
f = function(x) {c(-x[1]*(x[1]^2 - 1.1), -x[2]*(x[2]^2 - 1))}

# Sampling points
xs <- seq(-1.5, 1.5, length.out = 10)
ys <- seq(-1.5, 1.5, length.out = 15)

# Approximated potential
Vs <- approxPot2D(f, xs, ys, mode = 'horizontal')
}
