% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewsymm.R
\name{skewsymm}
\alias{skewsymm}
\title{Symmetric - skew decomposition of a square matrix}
\usage{
skewsymm(A)
}
\arguments{
\item{A}{the matrix to be decomposed}
}
\value{
B a list containing the symmetric and skew parts
}
\description{
Symmetric - skew decomposition of a square matrix
}
\examples{
A <- matrix(c(1,2,3,4), 2, 2)
B <- skewsymm(A)
}
